refactor: extract all components in separate files (SOC)
