+++ components/consent/context/hooks/index.ts
@@ -1,0 +1,36 @@
+import { useContext } from "react";
+import {
+  ConsentState,
+  ConsentManager,
+  ConsentDispatch,
+} from "../consent-manager";
+
+export function useConsent() {
+  const context = useContext(ConsentManager);
+  if (context === undefined) {
+    throw new Error(
+      "useGoogleTagManager must be used within a GoogleTagManagerProvider"
+    );
+  }
+  return context;
+}
+
+export function useConsentDispatch() {
+  const context = useContext(ConsentDispatch);
+  if (context === undefined) {
+    throw new Error(
+      "useGoogleTagManagerDispatch must be used within a GoogleTagManagerProvider"
+    );
+  }
+  return context;
+}
+
+export function useConsentState() {
+  const context = useContext(ConsentState);
+  if (context === undefined) {
+    throw new Error(
+      "useGoogleTagManagerConsent must be used within a GoogleTagManagerProvider"
+    );
+  }
+  return context;
+}
