+++ components/consent/context/index.ts
@@ -1,0 +1,2 @@
+export * from "./consent-manager"
+export * from "./hooks"
