+++ components/consent/index.ts
@@ -1,0 +1,4 @@
+export * from "./consent-manager";
+export * from "./context";
+export * from "./ui";
+export * from "./utils";
