+++ components/consent/ui/buttons/index.ts
@@ -1,0 +1,2 @@
+export * from './show-me-btn'
+export * from './triggers'
