+++ components/consent/ui/buttons/triggers.tsx
@@ -2,38 +2,11 @@
 
 import { Slot } from "@radix-ui/react-slot";
 
-import { Button, ButtonProps } from "../ui/button";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import { useGTM } from "@/components/cookies/context";
-import { background } from "./banner";
-import { BannerOptions } from "./banner-options";
+import { Button } from "../../../ui/button";
+import { ShowMeButton } from "./show-me-btn";
 
-import { cn } from "@/lib/utils";
-
-export function ShowMeButton({ btn, ...rest }: { btn?: ButtonProps }) {
-  const { tags } = useGTM();
-  return (
-    <Popover>
-      <PopoverTrigger asChild>
-        <Button {...btn} {...rest} onClick={console.log} />
-      </PopoverTrigger>
-      <PopoverContent
-        align="end"
-        side="top"
-        className={cn("relative, w-[500px] ", background)}
-      >
-        <div className="absolute rotate-[270deg] -left-[2.2rem] pt-3 top-[3.5rem] opacity-90 z-0 drop-shadow-md flex gap-2">
-          <p>Transparency</p>
-        </div>
-        <BannerOptions tags={tags} />
-      </PopoverContent>
-    </Popover>
-  );
-}
+import { useConsent, useConsentDispatch } from "../../context/hooks";
+import { convertTagsToCookies } from "../../utils";
 
 export function BannerTriggerGroup(props: ButtonGroupProps) {
   const ButtonGroupSlot = props.asChild ? Slot : BannerTriggers;
@@ -53,6 +26,8 @@ const isPro = !(process.env.NEXT_PUBLIC_FEATURE_PRO === "true");
 
 function BannerTriggers(props: React.PropsWithChildren<BannerTriggersProps>) {
   const { asChild, buttons, children, ...rest } = props;
+  const { handleConsentUpdate } = useConsentDispatch();
+  const { tags } = useConsent();
 
   let btns = buttons ?? _buttons;
   if (btns && btns.length > 2) {
@@ -71,7 +46,16 @@ function BannerTriggers(props: React.PropsWithChildren<BannerTriggersProps>) {
               // only show the feature button if the user has pro subscription
               return <ShowMeButton key={i} {...btn} />;
             }
-            return <Button key={i} {...btn} {...rest} />;
+            return (
+              <Button
+                key={i}
+                {...btn}
+                {...rest}
+                onClick={() => {
+                  handleConsentUpdate(convertTagsToCookies(tags));
+                }}
+              />
+            );
           })
         : null}
     </>
