+++ components/consent/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './banner'
+export * from './buttons'
+export * from './options'
