+++ components/consent/utils/handlers.ts
@@ -1,0 +1,6 @@
+import { toast } from "sonner";
+
+export const handlers = {
+  onSuccess: toast.success,
+  onError: toast.error,
+};
